Add data.js file and move data out of main.js file
